Add mtcp_create_context_on_lcore
This will create a context on the specified lcore and will not set the affinity of the thread (as we will launch it via eal_remote_launch and that should already set the affinity). As we do not set the affinity, you will have to be careful to launch application and mtcp thread on the same CPU or you will get race conditions.

This is very hacky, but unfortunately mTCP assumes that it can use all lcores, which is not the case in libmoon. And as a lot of stuff in mTCP uses the „index in a table“ instead of „pointer“ pattern, this seems to be the minimal invasive action.